Rename Commander to AbstractCommander anc BaseCommander to Commander
